Add radio field rendering

+++ resources/assets/js/main.js
@@ -1,7 +1,5 @@
-var formBuilder = new Vue({
-
-});
-
+Vue.config.debug = true;
+var s = require("underscore.string");
 
 var addFields = new Vue({
   el: '#add-fields',
@@ -25,10 +23,12 @@ methods: {
   },
     addEmptyField: function(){
         var blankField = {
-            id: 'unset',
+            id: Math.random().toString(36),
             name : '',
             description : '',
-            type: 'text'
+            type: 'text',
+            fieldOptions: [],
+            notSaved: true
         };
         addFields.fields = addFields.fields.concat(blankField);
     },
@@ -37,26 +37,76 @@ methods: {
         var field = _.findWhere(this.fields, {id: field_id});
         var fieldClass = '.field-' + field.id;
         var fieldURL = '/fields/' + field.id;
+        
 
-        $.ajax({
-          url: fieldURL,
-          type: 'DELETE',
-          success: function(result) {
-              $(fieldClass).fadeOut();
-          }
-        })
+        // Check for the "notSaved key"
+        if (typeof field.notSaved === 'undefined') {
+         $.ajax({
+            url: fieldURL,
+            type: 'DELETE',
+            success: function(result) {
+                $(fieldClass).fadeOut();
+            }
+          });          
+        }
+
+        // Drop the field from the array.
+        var fieldLocation = this.fields.indexOf(field);
+        this.fields.splice(fieldLocation, 1);
+
       } 
     },
     addFieldOption: function(field_id) {
         var blankFieldOption = {
-            id: 'unset',
+            id: Math.random().toString(36),
             name : '',
             text : '',
-            field_id: field_id
+            field_id: field_id,
+            notSaved: true
         };
         var field = _.findWhere(addFields.fields, {id: field_id}); 
         field.fieldOptions = field.fieldOptions.concat(blankFieldOption);      
+    },
+    deleteFieldOption: function(field_id, option_id) {
+      if (confirm('Are you sure you would like to delete this field?')) {
+        var field = _.findWhere(this.fields, {id: field_id});
+        var fieldOption = _.findWhere(field.fieldOptions, {id: option_id});
+        var fieldOptionClass = '.field-option-' + fieldOption.id;
+        var fieldOptionURL = '/fieldOptions/' + fieldOption.id;
+
+        // Check for the "notSaved key"
+        if (typeof fieldOption.notSaved === 'undefined') {
+          $.ajax({
+            url: fieldOptionURL,
+            type: 'DELETE',
+            success: function(result) {
+                $(fieldOptionClass).fadeOut();
+            }
+          });    
+        }
+
+         // Drop the field from the array.
+        var fieldOptionLocation = field.fieldOptions.indexOf(fieldOption);
+        console.log(field.fieldOptions);
+        field.fieldOptions.splice(fieldOptionLocation, 1);       
+
+      } 
+    },
+    // Check if a Field is elligible to have options
+    fieldHasOptions: function(field_id) {
+        var field = _.findWhere(this.fields, {id: field_id});
+        var fieldsWithOptions = this.form.field_types_with_options;
+        if ($.inArray(field.type, fieldsWithOptions) > -1) {
+          return true;
+        }
+        return false;
+    },
+
+    capitalizeString: function(string) {
+      var str = s(string).capitalize().value();
+      return str;
     }
+
 }
 });
